requirements.txt
# VirtualDisplayPy - Requirements
# GUI Framework
PySide6>=6.5.0
PyQt6>=6.5.0  # Alternative to PySide6

# Serial Communication
pyserial>=3.5
pyserial-asyncio>=0.6

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Plotting and Visualization
matplotlib>=3.5.0
plotly>=5.10.0

# Configuration Management
PyYAML>=6.0
configparser>=5.3.0
dataclasses-json>=0.5.7

# Networking
requests>=2.28.0
websockets>=10.4

# Testing
pytest>=7.2.0
pytest-qt>=4.2.0
pytest-asyncio>=0.21.0

# Logging and Monitoring
loguru>=0.6.0

# Utilities
click>=8.1.0  # For CLI
rich>=12.6.0  # For beautiful terminal output
typing-extensions>=4.4.0

# Development tools (optional)
black>=22.10.0
flake8>=5.0.0
mypy>=0.991